fix: address PRRT_kwDOSJAM6s5_Ic07 — extract shared helpers to _parity_utils.py
